Refactor generator into separate files per intent
